requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==3.7.1
flake8==2.2.3
django-nose==1.3
nose==1.3.4
